% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDef.R
\name{=}
\alias{=}
\title{Terser Function Declaration}
\usage{
left = right
}
\arguments{
\item{left}{A symbol or a function call.}

\item{right}{A R expression as body of the function.}
}
\value{
The assigned value or function.
}
\description{
A terse syntax of function declaration for people who are interested in
shooting yourself in the foot.
}
\details{
If left hand side is a symbol, it just do normal assignment like \code{<-}.

If left hand side is a function call, e.g. f(x, y), it creates a function named "f"
which has formal arguments of "x", "y" and function body defined by the right hand side,
in the environment calling it.
}
\examples{
x = 3
x

f(x, y) = x + y
f(2, 3)
identical(f, function(x, y) x + y)

g(x = 231) = log(x)
g()
identical(g, function(x = 231) log(x))

h(a, b = a^2) = a + b
h(1)
h(1, 2)
identical(h, function(a, b = a^2) a + b)

tan2(a) = sin(a)/cos(a)
tan2(pi)
}
